Update trivia.js
After running the game, I noticed that the answers were not being highlight according to the css so after some research I realized I had to target the classlist and use it's fuction to get the desire results. Also fixed a bug in the next question function that was only allowing a user to select the correct answer to select next question.